Use a much smaller value when offsetting to find grid intersections
Here we're calculating the nearest grid intersections from the source
of the ray. When we do that, we offset "inward" a bit so that we can
check the adjacent grid cells to see if they are a walls.

However, at close range and at certain angles, this offset caused the
distances to the horizontal and vertical grid boundaries to be slightly
wrong. This caused a visual distortion at the boundary between grid
cells on the wall, which looked like a small, flat protrusion.

By using a much smaller offset, we get the best of both worlds. We
still correctly "offset" to the adjacent grid cells, but the offset
isn't enough to cause visual abnormalities.