Annotate the <Scene> component
To make this easier, the map object has been converted to the Class
syntax, which Flow understands better. Also, note that I've provided
default `false` state for the `ArrowRight` state. That shouldn't be
necessary, but Flow thinks there's a type error without that. Note,
that `ArrowLeft` doesn't require this... just `ArrowRight`...